[limitation] Introduce Status & capture reset time
This change refactors the data that emitted by `Limiter`'s API to be
more actionable. In particular, three pieces of data are now returned as
part of a `Status` value struct:

* `limit` - the limit number, or max number or requests allowed
  per-period
* `remaining` - the positive number of requests remaining in the period,
  down to 0
* `reset_epoch_utc` - the "unix timestamp" or number of seconds since
  the UNIX epoch in UTC time. This gives a rough time when the period
  will end and a new period will begin

With these three pieces of information, we can eventually construct a
GitHub API style set of response headers such as:

* `X-RateLimit-Limit`
* `X-RateLimit-Remaining`
* `X-RateLimit-Reset`

Note that the time calculation uses the number of remaining seconds on
the Redis key and so there will be some rounding up and down a second
here or there when `reset_epoch_utc` is computed between different
requests. It shouldn't be a great concern and the lack of extra
persistence still leaves this as a great compromise.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>